pyproject.toml
[tool.poetry]
name = "counting-houses"
version = "0.1.0"
description = ""
authors = ["ClaraTump"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
matplotlib = "^3.9.1"
opencv-python = "^4.10.0.84"
torch = "^2.3.1"
torchvision = "^0.18.1"
yolov5 = "^7.0.13"
numpy = ">=1.20,<2.0"
pytest = "^8.2.2"
ultralytics = "^8.2.58"
pillow = "^10.4.0"
transformers = "^4.43.1"
datasets = "^2.20.0"
timm = "^1.0.7"
accelerate = "^0.33.0"
huggingface-hub = "^0.24.2"
albumentations = "^1.4.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
